requirements.txt
torch>=1.0
tensorboard
munch
opencv-python
h5py
sklearn